+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.2.0] - 2025-07-30
 
 ### Added 
+- Added Issue #1, Furling Sails Automation. Sheets are adjusted properly when sails furled
 - Added Issue #3, Text Feedback for Sail and Trim Commands (needs to be enabled in config)
 - Added Issue #4, Make Sail and Trim States Per-Boat Rather than Global
 - Added Issue #5, automatic jibe with gaffs, lateens, junksails (can be disabled in config, might not work on all ships)